requirements-dev.txt
-r requirements.txt
pytest==8.3.4
pytest-flask==1.3.0
coverage==7.6.9
factory-boy==3.3.1
